+++ physics/constants.h
@@ -1,4 +1,4 @@
 #pragma once
 
 double CONST_G = 0.00000000006673;
-float M2_APPROX_KG = 200
+double M2_APPROX_KG = 200
